backend/requirements.txt
praw>=7.7
anthropic>=0.40
fastapi>=0.115
uvicorn>=0.32
python-dotenv>=1.0
